added player position system
-added player position
-added player race position
-added player race position synching on network
-added player race position ui

+++ GeometryKart/Assets/Scripts/Checkpoint.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Checkpoint : MonoBehaviour
 {
-    [SerializeField] private byte checkpointId;
+    [SerializeField] private int checkpointId;
 
-    public byte CheckpointId => checkpointId;
+    public int CheckpointId => checkpointId;
 }
